refactor: move QC etc types into common crate
