+++ applications/tari_validator_node/src/p2p/proto/conversions/consensus.rs
@@ -26,21 +26,18 @@ use anyhow::anyhow;
 use tari_common_types::types::PublicKey;
 use tari_comms::types::CommsPublicKey;
 use tari_crypto::tari_utilities::ByteArray;
-use tari_dan_common_types::{ShardId, SubstateState};
-use tari_dan_core::models::{
-    vote_message::VoteMessage,
-    HotStuffMessage,
-    HotStuffTreeNode,
-    Node,
+use tari_dan_common_types::{
     ObjectPledge,
     QuorumCertificate,
     QuorumDecision,
     QuorumRejectReason,
+    ShardId,
     ShardVote,
-    TariDanPayload,
+    SubstateState,
     TreeNodeHash,
     ValidatorMetadata,
 };
+use tari_dan_core::models::{vote_message::VoteMessage, HotStuffMessage, HotStuffTreeNode, Node, TariDanPayload};
 use tari_engine_types::substate::Substate;
 
 use crate::p2p::proto;
