+++ applications/tari_validator_node/src/p2p/services/mempool/handle.rs
@@ -25,8 +25,7 @@ use std::{
     sync::{Arc, Mutex},
 };
 
-use tari_dan_common_types::ShardId;
-use tari_dan_core::models::TreeNodeHash;
+use tari_dan_common_types::{ShardId, TreeNodeHash};
 use tari_dan_engine::transaction::Transaction;
 use tari_template_lib::Hash;
 use tokio::sync::{broadcast, broadcast::error::RecvError, mpsc, mpsc::error::SendError};
