+++ clients/validator_node_client/src/types.rs
@@ -22,8 +22,13 @@
 
 use serde::{Deserialize, Serialize};
 use tari_common_types::types::{FixedHash, PublicKey};
-use tari_dan_common_types::{serde_with, Epoch, ShardId, SubstateChange};
-use tari_dan_core::models::{QuorumCertificate, QuorumDecision};
+use tari_dan_common_types::{
+    quorum_certificate::{QuorumCertificate, QuorumDecision},
+    serde_with,
+    Epoch,
+    ShardId,
+    SubstateChange,
+};
 use tari_engine_types::{
     commit_result::FinalizeResult,
     instruction::Instruction,
