+++ dan_layer/common_types/src/node_height.rs
@@ -1,0 +1,50 @@
+//   Copyright 2022 The Tari Project
+//   SPDX-License-Identifier: BSD-3-Clause
+
+use std::{
+    cmp::Ordering,
+    fmt::{Display, Formatter},
+    ops::Add,
+};
+
+use serde::{Deserialize, Serialize};
+use tari_bor::{borsh, Encode};
+
+#[derive(Copy, Clone, Debug, PartialEq, Eq, Hash, Deserialize, Serialize, Encode)]
+pub struct NodeHeight(pub u64);
+
+impl NodeHeight {
+    pub fn as_u64(self) -> u64 {
+        self.0
+    }
+
+    pub fn to_le_bytes(self) -> [u8; 8] {
+        self.0.to_le_bytes()
+    }
+}
+
+impl Add for NodeHeight {
+    type Output = NodeHeight;
+
+    fn add(self, rhs: Self) -> Self::Output {
+        NodeHeight(self.0 + rhs.0)
+    }
+}
+
+impl PartialOrd for NodeHeight {
+    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
+        self.0.partial_cmp(&other.0)
+    }
+}
+
+impl From<u64> for NodeHeight {
+    fn from(height: u64) -> Self {
+        NodeHeight(height)
+    }
+}
+
+impl Display for NodeHeight {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        write!(f, "NodeHeight({})", self.0)
+    }
+}
