+++ dan_layer/common_types/src/object_pledge.rs
@@ -1,0 +1,16 @@
+//   Copyright 2022 The Tari Project
+//   SPDX-License-Identifier: BSD-3-Clause
+
+use serde::{Deserialize, Serialize};
+use tari_bor::borsh::BorshSerialize;
+
+use crate::{NodeHeight, PayloadId, ShardId, SubstateState};
+
+#[derive(Debug, Clone, Deserialize, Serialize, BorshSerialize)]
+pub struct ObjectPledge {
+    pub shard_id: ShardId,
+    pub current_state: SubstateState,
+    // pub current_state_hash: FixedHash,
+    pub pledged_to_payload: PayloadId,
+    pub pledged_until: NodeHeight,
+}
