+++ dan_layer/common_types/src/shard_id.rs
@@ -1,0 +1,114 @@
+//   Copyright 2022 The Tari Project
+//   SPDX-License-Identifier: BSD-3-Clause
+
+use std::{
+    cmp::Ordering,
+    fmt,
+    fmt::{Display, Formatter},
+    str::FromStr,
+};
+
+use serde::{Deserialize, Serialize};
+use tari_bor::Encode;
+use tari_common_types::types::{FixedHash, FixedHashSizeError};
+use tari_engine_types::substate::SubstateAddress;
+use tari_utilities::hex::{from_hex, Hex};
+
+use crate::serde_with;
+
+#[derive(Copy, Clone, Debug, Eq, PartialEq, Hash, Serialize, Deserialize, Encode)]
+pub struct ShardId(#[serde(with = "serde_with::hex")] pub [u8; 32]);
+
+impl ShardId {
+    pub fn from_address(addr: &SubstateAddress) -> Self {
+        match addr {
+            SubstateAddress::Component(addr) => addr.into_array().into(),
+            SubstateAddress::Resource(addr) => addr.into_array().into(),
+            SubstateAddress::Vault(vault_id) => vault_id.into_array().into(),
+        }
+    }
+
+    pub fn new(id: FixedHash) -> Self {
+        let mut v = [0u8; 32];
+        v.copy_from_slice(id.as_slice());
+        Self(v)
+    }
+
+    pub fn as_bytes(&self) -> &[u8] {
+        &self.0
+    }
+
+    pub fn from_bytes(bytes: &[u8]) -> Result<Self, FixedHashSizeError> {
+        FixedHash::try_from(bytes).map(Self::new)
+    }
+
+    pub fn into_array(self) -> [u8; 32] {
+        self.0
+    }
+
+    pub fn zero() -> Self {
+        Self::new(FixedHash::default())
+    }
+}
+
+impl From<[u8; 32]> for ShardId {
+    fn from(bytes: [u8; 32]) -> Self {
+        Self(bytes)
+    }
+}
+
+impl From<ShardId> for Vec<u8> {
+    fn from(s: ShardId) -> Self {
+        s.as_bytes().to_vec()
+    }
+}
+
+impl TryFrom<Vec<u8>> for ShardId {
+    type Error = FixedHashSizeError;
+
+    fn try_from(value: Vec<u8>) -> Result<Self, Self::Error> {
+        Self::from_bytes(&value)
+    }
+}
+
+impl TryFrom<&[u8]> for ShardId {
+    type Error = FixedHashSizeError;
+
+    fn try_from(value: &[u8]) -> Result<Self, Self::Error> {
+        Self::from_bytes(value)
+    }
+}
+
+impl AsRef<[u8]> for ShardId {
+    fn as_ref(&self) -> &[u8] {
+        self.as_bytes()
+    }
+}
+
+impl PartialOrd for ShardId {
+    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
+        self.0.partial_cmp(&other.0)
+    }
+}
+
+impl Ord for ShardId {
+    fn cmp(&self, other: &Self) -> Ordering {
+        self.0.cmp(&other.0)
+    }
+}
+
+impl Display for ShardId {
+    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
+        write!(f, "{}", self.0.to_hex())
+    }
+}
+
+impl FromStr for ShardId {
+    type Err = FixedHashSizeError;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        // TODO: error isnt correct
+        let bytes = from_hex(s).map_err(|_| FixedHashSizeError)?;
+        Self::from_bytes(&bytes)
+    }
+}
