--- dan_layer/common_types/src/storage/mod.rs
@@ -1,6 +1,0 @@
-// Copyright 2022 The Tari Project
-// SPDX-License-Identifier: BSD-3-Clause
-
-mod unit_of_work_tracker;
-
-pub use unit_of_work_tracker::UnitOfWorkTracker;
