+++ dan_layer/core/src/digital_assets_error.rs
@@ -23,7 +23,6 @@
 use prost::DecodeError;
 use tari_common_types::types::FixedHashSizeError;
 use tari_comms_dht::outbound::DhtOutboundError;
-use tari_dan_engine::state::error::StateStorageError;
 use thiserror::Error;
 
 use crate::{models::ModelError, services::ValidatorNodeClientError, storage::StorageError};
@@ -95,8 +94,6 @@ pub enum DigitalAssetError {
     DecodeError(#[from] DecodeError),
     #[error("Invalid committee public key hex")]
     InvalidCommitteePublicKeyHex,
-    #[error("State storage error:{0}")]
-    StateStorageError(#[from] StateStorageError),
     #[error("Hash size error: {0}")]
     HashSizeError(#[from] FixedHashSizeError),
     #[error("Failed to register the validator node: {0}")]
