+++ dan_layer/core/src/services/leader_strategy.rs
@@ -23,9 +23,9 @@
 use digest::Digest;
 use tari_common_types::types::FixedHash;
 use tari_crypto::hash::blake2::Blake256;
-use tari_dan_common_types::{PayloadId, ShardId};
+use tari_dan_common_types::{NodeAddressable, PayloadId, ShardId};
 
-use crate::{models::Committee, services::infrastructure_services::NodeAddressable};
+use crate::models::Committee;
 
 pub trait LeaderStrategy<TAddr: NodeAddressable> {
     fn calculate_leader(&self, committee: &Committee<TAddr>, payload: PayloadId, shard: ShardId, round: u32) -> u32;
