+++ dan_layer/core/src/workers/events.rs
@@ -20,10 +20,9 @@
 //   WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE
 //   USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
+use tari_dan_common_types::QuorumCertificate;
 use tari_engine_types::commit_result::FinalizeResult;
 use tokio::sync::broadcast;
-
-use crate::models::QuorumCertificate;
 
 /// Wraps a broadcast sender, allowing a subscription (Receiver) to be obtained but removing the ability to send an
 /// event.
