+++ dan_layer/core/src/workers/hotstuff_error.rs
@@ -20,12 +20,11 @@
 //  WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE
 //  USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-use tari_dan_common_types::{Epoch, ShardId};
+use tari_dan_common_types::{Epoch, NodeHeight, ShardId};
 use tari_engine_types::commit_result::RejectReason;
 use thiserror::Error;
 
 use crate::{
-    models::NodeHeight,
     services::{epoch_manager::EpochManagerError, PayloadProcessorError},
     storage::{shard_store::StoreError, StorageError},
 };
